# Speed up YAML DB loading by removing ShowStatus