@aleos89 2023-02-10
I guess removing the count shouldn't be bad to loading as an error will show for the faulty entry!