@Lemongrass3110 2023-02-10
```
// windows with message
[Info]: Parsing [db/item_db.yml](ITEM_DB)...
[Info]: File Load to Buffer:            166

// windows nomsg
[Info]: Parsing [db/item_db.yml](ITEM_DB)...
[Info]: File Load to Buffer:            4621 <--------------------------------------------

// linux with message
[Info]: Parsing [db/item_db.yml](ITEM_DB)...
[Info]: File Load to Buffer:          20

// linux nomsg
[Info]: Parsing [db/item_db.yml](ITEM_DB)...
[Info]: File Load to Buffer:          19
```

I guess your machine had a hickup during your windows test without messages.
It is the only test where loading the file took significantly longer compared to all others.