@Lemongrass3110 2023-02-10
> nice, this reduce the startup time from around 20+ seconds to 10+ seconds.
> 
> and if i further remove the `Loading maps [xxx]` and `Loading NPCs .....` it further drop from 20+ seconds to 5+ seconds.
> 
> I think we should consider removing `ShowStatus` for `Loading xyz` infomation or make it into a config or define the config ? or just debug mode only ?

Every string preparation and console output that we reduce will reduce the loading time in general.