@vstumpf 2023-02-10
> I think we should consider removing `ShowStatus` for `Loading xyz` infomation
> or make it into a config or define the config ? or just debug mode only ?

I agree, we can move these to an `#ifdef DEBUG`

alternatively, we dump the entire showmsg and switch to a real logger, maybe [glog](https://github.com/google/glog)?