@Lemongrass3110 2023-02-10
> > I think we should consider removing `ShowStatus` for `Loading xyz` infomation
> 
> > or make it into a config or define the config ? or just debug mode only ?
> 
> 
> 
> I agree, we can move these to an `#ifdef DEBUG` in a future PR
> 
> 
> 
> alternatively, we dump the entire showmsg and switch to a real logger, maybe [glog](https://github.com/google/glog)?

I am talking about moving what you just removed into an ifdef and not about everything. Like if DEBUG use old variant, if not DEBUG then use your new logic.

And we can check if we want to replace the whole logging. But I guess we should check so that whatever we use then can also support not only logging to console, but also files/sql like we also have right now. So that we are able to harmonize all that.