@vstumpf 2023-02-10
> I am talking about moving what you just removed into an ifdef and not about everything. Like if DEBUG use old variant, if not DEBUG then use your new logic.

Yup, I went ahead and did that.

> And we can check if we want to replace the whole logging. But I guess we should check so that whatever we use then can also support not only logging to console, but also files/sql like we also have right now. So that we are able to harmonize all that.

I know glog can write to files as well as stdout/err, I don't think it can log to sql. that being said, it might make more sense to have another tool outside of the server process to read the output and add it to the sqldb.